% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{makeAbbrStemDict}
\alias{makeAbbrStemDict}
\title{makeAbbrStemDict}
\usage{
makeAbbrStemDict(term.vec, min.len = 3, min.share = NA)
}
\arguments{
\item{term.vec}{Character vector of input words}

\item{min.len}{The  length of word to stop of abbreviate searching}

\item{min.share}{The minimal ratio of lengths of two words to consider the first word as an abbreviation of the second word}
}
\value{
data frame with 3 columns: input words, intermediate stemming words, output  words
}
\description{
Creating a dictionary of words to transform  to their most short overall reduction (multi-core based on a basic parallel package)
}
\examples{
words <- c('Stemming','Stemmin','Stemm','Stem','Stemmi','SomeWord','St')
dict <- makeAbbrStemDict(term.vec=words,min.share=.8)
View(dict)
}
