% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{makeReducedSparceMatrix}
\alias{makeReducedSparceMatrix}
\title{makeReducedSparceMatrix}
\usage{
makeReducedSparceMatrix(
  dfmSparceMatrix,
  trainIndex,
  removePotentiallyEmptyRows = T
)
}
\arguments{
\item{dfmSparceMatrix}{Document-term (feature) matrix. May be sparse or traditional.}

\item{trainIndex}{True/False vector is the same length as the number of rows of the initial matrix. Truth - is record from training sample.}

\item{removePotentiallyEmptyRows}{A sign of whether it is necessary to cut the matrix on the empty rows (if they appeared after the column removal)}
}
\value{
Function return list with reduced matrix and reduced (if removePotentiallyEmptyRows==T) trainIndex for model prediction
}
\description{
Removal from an unnecessary column matrix (words) that exist in the training set and do not exist in the prediction set.
After that (optional): removal of empty rows that were formed after removing columns in the previous step (such rows do not carry predictive strength but occupy a place in memory)
}
