% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{makeSeparatedWords}
\alias{makeSeparatedWords}
\title{makeSeparatedWords}
\usage{
makeSeparatedWords(txt)
}
\arguments{
\item{txt}{Character vector of input texts}
}
\value{
Character vector of separated words
}
\description{
Separating merged words using letter case, language and digits
}
\examples{
products <- data.frame(life_case=c('"camel style"','forgotten CAPS LOCK','some words in upper','english~not english','digits and words'),
                       primer=c('ChateauBaronBellevueCoteDeBourgBordeaux', 'rIOJAcRIANZAbODEGASmARQUES', 'PINOTnoirBOURGOGNEdomaineMASSE', 'colombard ugniР±Р»Р°РЅles betesСЂРѕР·РѕРІРѕРµcomte Tolosan', 'Valencisco Rioja Blanco, Bodeguera de Valencisco0.75vol'))
products$splitted <- makeSeparatedWords(products$primer)
View(products)
}
