% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aboutDataSet_rawProducts.R
\docType{data}
\name{rawProducts}
\alias{rawProducts}
\title{rawProducts}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 419 rows and 2 columns.
}
\usage{
rawProducts
}
\description{
The dictionary of the original variations of the same names of wine related to the wine
in the standards of the company names through the key field "Standartid".
As part of this example, this is a training data set (and validation data set)
}
\examples{
data(standardProducts)
data(rawProducts)
mergeData<-merge(standardProducts,rawProducts)
View(mergeData)
}
\keyword{datasets}
