% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aboutDataSet_StandartProducts.R
\docType{data}
\name{standardProducts}
\alias{standardProducts}
\title{standardProducts}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 120 rows and 3 columns.
}
\usage{
standardProducts
}
\description{
Wines dictionary in the name of the company standards.
As an example, each row of this table is a class description
to which many variations of the original wine names
belong (by "standartId" in the "rawProducts" table)
}
\examples{
data(standardProducts)
data(rawProducts)
mergeData<-merge(standardProducts,rawProducts)
View(mergeData)
}
\keyword{datasets}
